KRACOEUS-6956 : move ProposalDevelopmentControllerBase.java  to propdev core.
