+++ coeus-code/src/main/java/org/kuali/coeus/propdev/impl/basic/ProposalDevelopmentHomeController.java
@@ -17,12 +17,12 @@ package org.kuali.coeus.propdev.impl.basic;
 
 import org.kuali.coeus.common.framework.sponsor.Sponsor;
 import org.kuali.coeus.common.specialreview.impl.rule.event.SaveDocumentSpecialReviewEvent;
+import org.kuali.coeus.propdev.impl.core.ProposalDevelopmentControllerBase;
 import org.kuali.coeus.sys.framework.controller.UifExportControllerService;
 import org.kuali.kra.bo.ScienceKeyword;
 import org.kuali.kra.proposaldevelopment.bo.PropScienceKeyword;
 import org.kuali.kra.proposaldevelopment.document.ProposalDevelopmentDocument;
 import org.kuali.kra.proposaldevelopment.specialreview.ProposalSpecialReview;
-import org.kuali.kra.proposaldevelopment.web.krad.ProposalDevelopmentControllerBase;
 import org.kuali.kra.proposaldevelopment.web.krad.ProposalDevelopmentDocumentForm;
 import org.kuali.rice.core.api.criteria.QueryByCriteria;
 import org.kuali.rice.krad.exception.AuthorizationException;
