+++ coeus-code/src/main/java/org/kuali/kra/proposaldevelopment/web/krad/ProposalDevelopmentPersonnelController.java
@@ -18,6 +18,7 @@ package org.kuali.kra.proposaldevelopment.web.krad;
 import org.apache.commons.lang3.StringUtils;
 import org.kuali.coeus.common.framework.person.KcPerson;
 import org.kuali.coeus.common.framework.rolodex.Rolodex;
+import org.kuali.coeus.propdev.impl.core.ProposalDevelopmentControllerBase;
 import org.kuali.kra.proposaldevelopment.bo.ProposalPerson;
 import org.kuali.kra.proposaldevelopment.document.ProposalDevelopmentDocument;
 import org.kuali.kra.proposaldevelopment.service.KeyPersonnelService;
