+++ coeus-code/src/main/java/org/kuali/kra/proposaldevelopment/web/krad/ProposalDevelopmentQuestionnaireController.java
@@ -18,6 +18,7 @@ package org.kuali.kra.proposaldevelopment.web.krad;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import org.kuali.coeus.propdev.impl.core.ProposalDevelopmentControllerBase;
 import org.kuali.kra.proposaldevelopment.document.ProposalDevelopmentDocument;
 import org.kuali.kra.proposaldevelopment.specialreview.ProposalSpecialReview;
 import org.kuali.rice.krad.web.form.DocumentFormBase;
