+++ coeus-code/src/main/java/org/kuali/kra/proposaldevelopment/web/krad/ProposalDevelopmentS2SController.java
@@ -16,6 +16,7 @@
 package org.kuali.kra.proposaldevelopment.web.krad;
 
 import org.apache.commons.lang3.StringUtils;
+import org.kuali.coeus.propdev.impl.core.ProposalDevelopmentControllerBase;
 import org.kuali.coeus.sys.framework.service.KcServiceLocator;
 import org.kuali.kra.infrastructure.Constants;
 import org.kuali.kra.infrastructure.KeyConstants;
